Addition of radio-astronomy use-case (#358)
* Add radio-astronomy use case

In use-cases/radio-astronomy, an itwinai integration of ML pulsar detection workflow developed by HTW Berlin is presented. This use-case features a sophisticated ML pipeline composed of 3 models. This workflow has been tested with torch-ddp, microsoft deepspeed and horovod distributed learning strategies. The batch-jsc.sh script provides needed functionality for each strategy, so the user only needs to modify the config.yaml file to change the strategy.

* Test-suite

The use-case also has a test suite which supports automatic generation of synthetic data in the temporary folder and trains the models on this synthetic data. 

* Pulsar-plugin https://github.com/interTwin-eu/pulsar-plugin

Besides this use-case, which is kept in itwinai repository for demonstration reasons, a plug-in solution for itwinai was developed for ease of modification of the integration by HTW Berlin developers, if necessary. The data.py and trainer.py files, which provide the integration between use-case source code and itwinai code, are very similar between this use-case and the plug-in as of May 2025. 

* Running from config.yaml file

Five pipelines, one for data generation, three for individual model training and one for full workflow evaluation, are set-up in config.yaml file.

- Alex Krochak, FZJ