Fix radio tests (#433)
I've turned off the ray dashboard for the synthetic data generation. In the container, when running the pytest pipeline for the use-cases, it would attempt to create a ray dashboard and crash Ray without killing the job. Disabling the dashboard fixes the problem for me locally (Macbook) and on Juwels Booster, but the container likely doesn't have enough RAM to execute the job either way. So I'll leave the tests disabled, but look into an alternative GH testing pipeline that can be executed on Juwels Booster with sufficient resources.